requirements.txt
appdirs==1.4.4
asgiref==3.5.2
astroid==2.12.5
autopep8==1.7.0
Babel==2.10.3
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
cryptography==2.8
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
Django==3.2.15
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.16
filelock==3.8.0
flake8==5.0.4
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.4.3
mapview==1.0.5
Markdown==3.1.1
MarkupSafe==2.0.1
mccabe==0.6.1
mock==3.0.5
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.10.1
peewee==3.14.0
phonenumbers==8.11.1
Pillow==7.0.0
pipenv==2020.11.15
platformdirs==2.5.2
psycopg2==2.8.4
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.9.4
pyflakes==2.5.0
Pygments==2.7.3
PyJWT==2.4.0
pylint==2.6.0
pypiwin32==223
python3-openid==3.2.0
pytz==2019.3
pywin32==300
rave-python==1.1.2
requests==2.22.0
requests-oauthlib==1.3.1
six==1.13.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.4.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.4.0
